color-detection/requirements.txt
opencv-python==4.6.0.66
numpy==1.24.3
pillow==9.2.0